Adding new annotation parsing method
