+++ SourceryFramework/Sources/Parsing/Utils/AnnotationsParser.swift
@@ -18,7 +18,7 @@ public struct AnnotationsParser {
         case file(Annotations)
     }
 
-    private struct Line {
+    struct Line {
         enum LineType {
             case propertyWrapper
             case macros
@@ -37,7 +37,7 @@ public struct AnnotationsParser {
         let blockAnnotations: Annotations
     }
 
-    private let lines: [Line]
+    let lines: [AnnotationsParser.Line]
     private let contents: String
     private var parseDocumentation: Bool
     internal var sourceLocationConverter: SourceLocationConverter
@@ -86,7 +86,7 @@ public struct AnnotationsParser {
 
     func documentation(fromToken token: SyntaxProtocol) -> Documentation {
         guard parseDocumentation else {
-            return  []
+            return []
         }
         return documentationFrom(
           location: findLocation(syntax: token)
@@ -95,6 +95,33 @@ public struct AnnotationsParser {
 
     private func findLocation(syntax: SyntaxProtocol) -> SwiftSyntax.SourceLocation {
         return sourceLocationConverter.location(for: syntax.positionAfterSkippingLeadingTrivia)
+    }
+
+    private func from(location: SwiftSyntax.SourceLocation, syntax: SyntaxProtocol) -> Annotations {
+        var stop = false
+        var annotations = inlineFrom(line: (location.line, location.column), stop: &stop)
+        guard !stop else { return annotations }
+
+        let reversedArray = lines[0..<location.line-1].reversed()
+        for line in reversedArray {
+            line.annotations.forEach { annotation in
+                AnnotationsParser.append(key: annotation.key, value: annotation.value, to: &annotations)
+            }
+
+            if line.type != .comment
+                && line.type != .documentationComment
+                && line.type != .macros
+                && line.type != .propertyWrapper
+            {
+                break
+            }
+        }
+
+        lines[location.line-1].annotations.forEach { annotation in
+            AnnotationsParser.append(key: annotation.key, value: annotation.value, to: &annotations)
+        }
+
+        return annotations
     }
 
     private func from(location: SwiftSyntax.SourceLocation) -> Annotations {
@@ -108,7 +135,7 @@ public struct AnnotationsParser {
                 AnnotationsParser.append(key: annotation.key, value: annotation.value, to: &annotations)
             }
 
-            if line.type != .comment 
+            if line.type != .comment
                 && line.type != .documentationComment
                 && line.type != .macros
                 && line.type != .propertyWrapper
@@ -117,7 +144,7 @@ public struct AnnotationsParser {
             }
         }
 
-      lines[location.line-1].annotations.forEach { annotation in
+        lines[location.line-1].annotations.forEach { annotation in
             AnnotationsParser.append(key: annotation.key, value: annotation.value, to: &annotations)
         }
 
@@ -156,62 +183,63 @@ public struct AnnotationsParser {
         return documentation.reversed()
     }
 
-    func inlineFrom(line lineInfo: (line: Int, character: Int), stop: inout Bool) -> Annotations {
-        let sourceLine = lines[lineInfo.line - 1]
-        let utf8View = sourceLine.content.utf8
-        let startIndex = utf8View.startIndex
-        let endIndex = utf8View.index(startIndex, offsetBy: (lineInfo.character - 1))
-        let utf8Slice = utf8View[startIndex ..< endIndex]
-        let relevantContent = String(decoding: utf8Slice, as: UTF8.self)
-        var prefix = relevantContent.trimmingCharacters(in: .whitespaces)
-
-        guard !prefix.isEmpty else { return [:] }
-        var annotations = sourceLine.blockAnnotations // get block annotations for this line
-        sourceLine.annotations.forEach { annotation in  // TODO: verify
-            AnnotationsParser.append(key: annotation.key, value: annotation.value, to: &annotations)
-        }
-
-        // `case` is not included in the key of enum case definition, so we strip it manually
-        let isInsideCaseDefinition = prefix.trimmingCharacters(in: .whitespacesAndNewlines).hasPrefix("case")
-        prefix = prefix.trimmingPrefix("case").trimmingCharacters(in: .whitespaces)
-        var inlineCommentFound = false
-
-        while !prefix.isEmpty {
-            guard prefix.hasSuffix("*/"), let commentStart = prefix.range(of: "/*", options: [.backwards]) else {
-                break
-            }
-
-            inlineCommentFound = true
-
-            let comment = String(prefix[commentStart.lowerBound...])
-            for annotation in AnnotationsParser.parse(contents: comment)[0].annotations {
-                AnnotationsParser.append(key: annotation.key, value: annotation.value, to: &annotations)
-            }
-            prefix = prefix[..<commentStart.lowerBound].trimmingCharacters(in: .whitespaces)
-        }
-
-        if (inlineCommentFound || isInsideCaseDefinition) && !prefix.isEmpty {
-            stop = true
-            return annotations
-        }
-
-        // if previous line is not comment or has some trailing non-comment blocks
-        // we return currently aggregated annotations
-        // as annotations on previous line belong to previous declaration
-        if lineInfo.line - 2 > 0 {
-            let previousLine = lines[lineInfo.line - 2]
-            let content = previousLine.content.trimmingCharacters(in: .whitespaces)
-            
-            guard previousLine.type == .comment || previousLine.type == .documentationComment || previousLine.type == .propertyWrapper || previousLine.type == .macros, content.hasPrefix("//") || content.hasSuffix("*/") || content.hasPrefix("@") || content.hasPrefix("#") else {
-                stop = true
-                return annotations
-            }
-        }
-
-        return annotations
+    func inlineFrom(line lineInfo: (line: Int, character: Int), node: Syntax, stop: inout Bool) -> Annotations {
+//        let sourceLine = lines[lineInfo.line - 1]
+//        let utf8View = sourceLine.content.utf8
+//        let startIndex = utf8View.startIndex
+//        let endIndex = utf8View.index(startIndex, offsetBy: (lineInfo.character - 1))
+//        let utf8Slice = utf8View[startIndex ..< endIndex]
+//        let relevantContent = String(decoding: utf8Slice, as: UTF8.self)
+//        var prefix = relevantContent.trimmingCharacters(in: .whitespaces)
+//
+//        guard !prefix.isEmpty else { return [:] }
+//        var annotations = sourceLine.blockAnnotations // get block annotations for this line
+//        sourceLine.annotations.forEach { annotation in  // TODO: verify
+//            AnnotationsParser.append(key: annotation.key, value: annotation.value, to: &annotations)
+//        }
+//
+//        // `case` is not included in the key of enum case definition, so we strip it manually
+//        let isInsideCaseDefinition = prefix.trimmingCharacters(in: .whitespacesAndNewlines).hasPrefix("case")
+//        prefix = prefix.trimmingPrefix("case").trimmingCharacters(in: .whitespaces)
+//        var inlineCommentFound = false
+//
+//        while !prefix.isEmpty {
+//            guard prefix.hasSuffix("*/"), let commentStart = prefix.range(of: "/*", options: [.backwards]) else {
+//                break
+//            }
+//
+//            inlineCommentFound = true
+//
+//            let comment = String(prefix[commentStart.lowerBound...])
+//            for annotation in AnnotationsParser.parse(contents: comment)[0].annotations {
+//                AnnotationsParser.append(key: annotation.key, value: annotation.value, to: &annotations)
+//            }
+//            prefix = prefix[..<commentStart.lowerBound].trimmingCharacters(in: .whitespaces)
+//        }
+//
+//        if (inlineCommentFound || isInsideCaseDefinition) && !prefix.isEmpty {
+//            stop = true
+//            return annotations
+//        }
+//
+//        // if previous line is not comment or has some trailing non-comment blocks
+//        // we return currently aggregated annotations
+//        // as annotations on previous line belong to previous declaration
+//        if lineInfo.line - 2 > 0 {
+//            let previousLine = lines[lineInfo.line - 2]
+//            let content = previousLine.content.trimmingCharacters(in: .whitespaces)
+//            
+//            guard previousLine.type == .comment || previousLine.type == .documentationComment || previousLine.type == .propertyWrapper || previousLine.type == .macros, content.hasPrefix("//") || content.hasSuffix("*/") || content.hasPrefix("@") || content.hasPrefix("#") else {
+//                stop = true
+//                return annotations
+//            }
+//        }
+//
+//        return annotations
+        return [:]
     }
 
-    private static func parse(contents: String) -> [Line] {
+    static func parse(contents: String) -> [Line] {
         var annotationsBlock: Annotations?
         var fileAnnotationsBlock = Annotations()
